\name{EMT-package}
\alias{EMT-package}
\alias{EMT}
\docType{package}
\title{Exact Multinomial Test: Goodness-of-Fit Test for Discrete Multivariate Data}

\description{
  The package provides functions to carry out a Goodness-of-fit test for discrete multivariate data. 
  It is tested if a given observation is likely to have occurred under the assumption of an ab-initio model. 
  A p-value can be calculated using different distance measures between observed and expected frequencies. 
  A Monte Carlo method is provided to make the package capable of solving high-dimensional problems.  
  The main user functions are \code{multinomial.test} and \code{plotMultinom}.
}

\details{
\tabular{ll}{
Package: \tab CCP\cr
Type: \tab Package\cr
Version: \tab 1.3\cr
Date: \tab 2013-02-06\cr
License: \tab GPL\cr
}
}

\author{Uwe Menzel

  Maintainer: Uwe Menzel  <uwemenzel@gmail.com>
}

\keyword{htest}
\keyword{multivariate}


