% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{emt.feedback.no_score}
\alias{emt.feedback.no_score}
\title{EMT feedback (no score)}
\usage{
emt.feedback.no_score(dict = emt::emt_dict)
}
\arguments{
\item{dict}{The psychTestR dictionary used for internationalisation.}
}
\description{
Here the participant is given no feedback at the end of the test.
}
\examples{
\dontrun{
demo_emt(feedback = emt.feedback.no_score())}
}
