% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{standalone_emt}
\alias{standalone_emt}
\title{Standalone EMT}
\usage{
standalone_emt(title = "Emotion matching test",
  admin_password = "replace-with-secure-password",
  researcher_email = NULL, languages = emt_languages(),
  dict = emt::emt_dict, ...)
}
\arguments{
\item{title}{(Scalar character) Title to display during testing.}

\item{admin_password}{(Scalar character) Password for accessing the admin panel.}

\item{researcher_email}{(Scalar character)
If not \code{NULL}, this researcher's email address is displayed
at the bottom of the screen so that online participants can ask for help.}

\item{languages}{(Character vector)
Determines the languages available to participants.
Currently the only possible language is Engish (\code{"EN"}).}

\item{dict}{The psychTestR dictionary used for internationalisation.}

\item{...}{Further arguments to be passed to \code{\link{emt}()}.}
}
\description{
This function launches a standalone testing session for the EMT.
This can be used for data collection, either in the laboratory or online.
}
